github/update-repo-secrets/requirements.txt
PyNaCl==1.5.0
requests==2.32.4
urllib3==2.5.0
